requirements.txt
hydra-core==1.3.2
matplotlib==3.5.1
mayavi==4.8.2
natsort==8.4.0
numpy==1.24.3
omegaconf==2.3.0
open3d==0.18.0
openTSNE==1.0.1
Pillow==10.3.0
pygame==2.5.2
pyparsing==2.4.7
pyrealsense2==2.55.1.6486
pyserial==3.5
PyYAML==6.0.1
scikit_learn==1.5.0
scipy==1.8.0
seaborn==0.13.2
setuptools==59.6.0
skimage==0.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.1
wandb==0.17.1
